+++ vector/src/main/java/im/vector/app/core/services/CallRingPlayer.kt
@@ -21,6 +21,7 @@ import android.media.AudioAttributes
 import android.media.AudioManager
 import android.media.MediaPlayer
 import android.os.Build
+import androidx.core.content.getSystemService
 import im.vector.app.R
 import timber.log.Timber
 
@@ -33,7 +34,7 @@ class CallRingPlayer(
     private var player: MediaPlayer? = null
 
     fun start() {
-        val audioManager: AudioManager = applicationContext.getSystemService(Context.AUDIO_SERVICE) as AudioManager
+        val audioManager = applicationContext.getSystemService<AudioManager>()!!
         player?.release()
         player = createPlayer()